huggingface-rag-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sentence-transformers==3.3.1
faiss-cpu==1.9.0
google-generativeai==0.8.3
numpy==1.26.4
pydantic==2.10.0
python-dotenv==1.0.1
